extract logic to external modules
